router contract introduced
